refactor dpm pm quality summary invocation routes
